+++ NEWS.md
@@ -1,3 +1,8 @@
+# filecacher 0.2.6
+
+* Another fix to `test-cached_read` to pass CRAN Package Check. 
+  Copied test data to a temporary directory to perform tests there.
+
 # filecacher 0.2.5
 
 * Fix to `test-cached_read` to pass CRAN Package Check.